Reintroduce keys for closing braces
The initial idea of the double-tap keys was that combining with Shift
would result in the ending brace. However, the planck still sents the
Shift modifier to the OS. For `]`, this results in Shift+], which is
translated to `}` by the OS.

Reintroducing the closing braces key allows me to figure out how to
avoid this behavior consistently.