Introduce the gaming layer
All tab-hold configurations on the keyboard have broken using the
keyboard for games. Holding the A key no longer dispatches holding the
left key to the computer, but instead dispatches holding the Alt key.

The gaming layer is a copy of the base layer, with all special hold
configurations removed. It is enabled as a new default layer, and allows
you to toggle back to the normal qwerty layer when you're done gaming.
The keyboard backlights indicate whether or not the gaming layer is
activated.