Add very naive wide layout
I've always been wanting to try moving outside columns to the middle,
moving the hands further from eachother. After attemping to design my
keymap to work with this wide layout for a long time, the idea sort of
died down.

This is my attempt at restarting this experiment, simply by moving these
columns in the middle and seeing what odd things I get around. I'm
expecting a ton of tweaks in the coming weeks from this way to naive attempt.